requirements.txt
boto3==1.14.20
botocore==1.17.20
certifi==2020.6.20
chardet==3.0.4
coloredlogs==14.0
decorator==4.4.2
docutils==0.15.2
humanfriendly==8.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
jmespath==0.10.0
moviepy==1.0.3
numpy==1.19.0
Pillow==7.2.0
proglog==0.1.9
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.14.0
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
tqdm==4.47.0
urllib3==1.25.9
